+++ app/src/main/kotlin/com/wire/android/ui/home/conversationslist/model/AvailabilityStatus.kt
@@ -1,0 +1,7 @@
+package com.wire.android.ui.home.conversationslist.model
+
+
+enum class AvailabilityStatus {
+    Available, Busy, Away, None
+}
+
