+++ app/src/main/kotlin/com/wire/android/ui/home/conversationslist/model/EventType.kt
@@ -1,0 +1,8 @@
+package com.wire.android.ui.home.conversationslist.model
+
+sealed class EventType {
+    data class UnreadMessage(val unreadMessageCount: Int) : EventType()
+    object UnreadMention : EventType()
+    object UnreadReply : EventType()
+    object MissedCall : EventType()
+}
