+++ app/src/main/kotlin/com/wire/android/ui/home/conversationslist/model/Membership.kt
@@ -1,0 +1,8 @@
+package com.wire.android.ui.home.conversationslist.model
+
+import androidx.annotation.StringRes
+import com.wire.android.R
+
+enum class Membership(@StringRes val stringResourceId: Int) {
+    Guest(R.string.label_membership_guest), External(R.string.label_memebership_external), None(-1)
+}
