+++ app/src/main/kotlin/com/wire/android/ui/home/conversationslist/model/NewActivity.kt
@@ -5,9 +5,3 @@ data class NewActivity(
     val conversation: Conversation,
 )
 
-sealed class EventType {
-    data class UnreadMessage(val unreadMessageCount: Int) : EventType()
-    object UnreadMention : EventType()
-    object UnreadReply : EventType()
-    object MissedCall : EventType()
-}
